Update Solution.cs
Would this be a simpler solution to the problem? I believe this is also an O(1) solution (but I am extremely new to this field, so may be way off base). This is also my first contribution to a project, so I am sorry if it is not done in the proper way.